new branch

+++ src/main/java/org/example/Main.java
@@ -12,12 +12,12 @@ public class Main {
 
         // Range : -1000 ; 1000
         for (int i = 0; i < iterations; i++) {
-            float x = random.nextFloat() * 2e3f - 1e3f;
-            float y = random.nextFloat() * 2e3f - 1e3f;
-            float z = random.nextFloat() * 2e3f - 1e3f;
+            double x = random.nextFloat() * 2e3f - 1e3f;
+            double y = random.nextFloat() * 2e3f - 1e3f;
+            double z = random.nextFloat() * 2e3f - 1e3f;
 
-            float leftSide = (x + y) + z;
-            float rightSide = x + (y + z);
+            double leftSide = (x + y) + z;
+            double rightSide = x + (y + z);
 
             if (leftSide - rightSide == 0) {
                 validCount++;
